mode cloudwatch adapter functions into adapter.go
